@@ -4,7 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve --host=0.0.0.0",
-    "build": "ng build --prod",
+    "build": "ng build --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
@@ -13,31 +13,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.2.14",
-    "@angular/common": "^11.2.14",
-    "@angular/compiler": "^11.2.14",
-    "@angular/core": "^11.2.14",
-    "@angular/forms": "^11.2.14",
-    "@angular/localize": "^11.2.14",
-    "@angular/platform-browser": "^11.2.14",
-    "@angular/platform-browser-dynamic": "^11.2.14",
-    "@angular/router": "^11.2.14",
+    "@angular/animations": "^12.0.5",
+    "@angular/common": "^12.0.5",
+    "@angular/compiler": "^12.0.5",
+    "@angular/core": "^12.0.5",
+    "@angular/forms": "^12.0.5",
+    "@angular/localize": "^12.0.5",
+    "@angular/platform-browser": "^12.0.5",
+    "@angular/platform-browser-dynamic": "^12.0.5",
+    "@angular/router": "^12.0.5",
     "@sweetalert2/ngx-sweetalert2": "^10.0.0",
-    "angular-datatables": "^11.2.0",
+    "angular-datatables": "^12.0.0",
     "datatables.net": "^1.10.25",
     "datatables.net-dt": "^1.10.25",
     "jquery": "^3.6.0",
-    "rxjs": "^6.5.3",
+    "rxjs": "^7.1.0",
     "sweetalert2": "^11.0.18",
     "tslib": "^2.3.0",
     "uuid": "^8.3.2",
-    "zone.js": "^0.11.4"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.13",
-    "@angular/cli": "^11.2.13",
-    "@angular/compiler-cli": "^11.2.14",
-    "@angular/language-service": "^11.2.14",
+    "@angular-devkit/build-angular": "^12.0.5",
+    "@angular/cli": "^12.0.5",
+    "@angular/compiler-cli": "^12.0.5",
+    "@angular/language-service": "^12.0.5",
     "@compodoc/compodoc": "^1.1.11",
     "@types/datatables.net": "^1.10.19",
     "@types/jasmine": "^3.7.7",
@@ -53,8 +53,8 @@
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.6.0",
     "protractor": "~7.0.0",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.0.0",
     "tslint": "~6.1.3",
-    "typescript": "^4.1.5"
+    "typescript": "^4.2.4"
   }
 }
