@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { HttpClient, HttpHeaders } from "@angular/common/http";
 import { environment } from "../../environments/environment";
+import { lastValueFrom } from "rxjs";
 @Injectable({
   providedIn: "root",
 })
@@ -19,22 +20,22 @@ export class ExampleService {
   };
 
   getUsers(): Promise<any> {
-    return this.http.get(this.BASEURL + "example/users", this.httpOptions).toPromise();
+    return lastValueFrom(this.http.get(this.BASEURL + "example/users", this.httpOptions));
   }
 
   getUsersById(id: number): Promise<any> {
-    return this.http.get(this.BASEURL + "example/users/" + id, this.httpOptions).toPromise();
+    return lastValueFrom(this.http.get(this.BASEURL + "example/users/" + id, this.httpOptions));
   }
 
   deleteUsersById(id: number): Promise<any> {
-    return this.http.delete(this.BASEURL + "example/users?id=" + id, this.httpOptions).toPromise();
+    return lastValueFrom(this.http.delete(this.BASEURL + "example/users?id=" + id, this.httpOptions));
   }
 
   postUsers(login: string, password: string, email: string): Promise<any> {
-    return this.http.post(this.BASEURL + "example/users", {login, password, email}, this.httpOptions).toPromise();
+    return lastValueFrom(this.http.post(this.BASEURL + "example/users", {login, password, email}, this.httpOptions));
   }
 
   putUsers(id: number, login: string, password: string, email: string): Promise<any> {
-    return this.http.put(this.BASEURL + "example/users", {id, login, password, email}, this.httpOptions).toPromise();
+    return lastValueFrom(this.http.put(this.BASEURL + "example/users", {id, login, password, email}, this.httpOptions));
   }
 }
